Rounds tile_lat and tile_lon indices.

+++ src/viewport_entropy_toolkit/utilities/entropy_utils.py
@@ -410,10 +410,10 @@ def find_naive_tile_index(
         Dict[Vector, float]: Dictionary mapping tile centers to weights.
     """
     # calculate the tile this vector sits in.
-    tile_lon_index = point.lon / tile_width
+    tile_lon_index = int(point.lon / tile_width)
     if (point.lon < 0):
         tile_lon_index -= 1
-    tile_lat_index = point.lat / tile_height
+    tile_lat_index = int(point.lat / tile_height)
     if (point.lat < 0):
         tile_lat_index += 1
